Refactor: made custom hook + changed product page according to it
